Implement a coolc format ast dumper
Had to make some changes to ast types to reflect the reference structure
more closely.

I'm quite fond of the solution with Visitor, DefaultVisitor, and a
blanket impl of the former for the implementors of the latter.
Not yet sure how much hassle that's going to prove to be, though.