Implement a local allocator
A month ago when I was thinking how to approach name resolution, I've
decided against making any name maps believing it would not be
necessary.
Therefore, the allocator is going to allocate locals on-the-fly while
lowering the ast to wasm.
After a method is processed, it'll have a complete list of all needed
locals and their types.