Set up parser tests
There is a set of reference Cool compiler tools (lexer, parser, etc)
which generate output in a custom format (referred to as the coolc format
in the code).
I've implemented the support for this format a long time ago and was
manually validating my output against the reference the entire time...
well, not exactly manually — I had a script to do that,
but it wasn't staged.

This commit introduces basic infrastructure to perform parser testing.
Each test file is parsed, and the result is compared against the
reference.
However, the coolc format is very basic — it doesn't include span
information, for instance.
There are also some small discrepancies between the reference compiler's
ast representation and mine.
To accomodate for that, I use the `insta` crate to perform snapshot
testing as well.

The `insta` crate comes with a `cargo-insta` extension command which
provides a nice interface for reviewing snapshot changes.
It has to be installed separately:

```
$ cargo install cargo-insta
```

I've also added `Makefile`s to generate reference files.
I'm including these in the git tree so the makefile is not necessary for
running the existing tests.

Lastly, a couple of known issues:

1. I couldn't find a way to have the snapshots directory reside in
   `tests/parser`, so gotta live with that.
   At least the file names start with `parser__`, quenching any fear of
   ambiguity.

2. I've skimmed through the generated snapshots to make sure I'm getting
   something along the lines of what I expect, but I didn't thoroughly
   validate them (spans may be a bit off, for instance).

3. Right now I can only add parser tests.
   The plan is to extend the testing system to cover typeck and codegen.
   Actually, the latter was the main reason to introduce this in the
   first place — running programs in Chromium manually is fun but
   definitely not scalable.
   Moreover, it's not even my main browser.

   I'm thinking of using nodejs as a substitute for Chromium.
   It provides the necessary wasm API and seems to support the
   `--experimental-wasm-gc` V8 command-line flag.

4. The test coverage is pretty low.
   I'm not going for a 100% here, since that requires quite a lot of
   effort, to be honest.
   In fact, it seems to me that codegen testing yields the greatest
   benefit for the same amount of effort, since it's dependent on all
   the previous stages working correctly,
   so that'll be my priority for now.