Switch to a wast fork
The GC proposal has progressed quite a bit, and the support for it in
wast has lagged behind heavily. Binaryen no longer accepts the
deprecated instruction encodings — and figuring this out it took me much
longer than what I would like to admit (I was seriously considering
patching its parser so that it would recognize the output of wacc).

This commit switches to a wast fork I've made to bring in the
instructions I need for my compiler. Hopefully I'll be able to go back
to the upstream version soon enough.